services/recommendation-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pymongo==4.6.0
structlog==23.2.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Recommendation Engine specific dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4
ortools==9.8.3296
